+++ lib/celluloid.rb
@@ -1,7 +1,9 @@
 require 'logger'
 require 'thread'
+require 'timeout'
 
 module Celluloid
+  SHUTDOWN_TIMEOUT = 60 # How long actors have to terminate
   @logger = Logger.new STDERR
 
   class << self
@@ -247,6 +249,25 @@ module Celluloid
     end
 
     super
+  end
+
+  at_exit do
+    # Give all actors one minute to terminate
+    Timeout.timeout(SHUTDOWN_TIMEOUT) do
+      futures = Celluloid::Actor.all.each do |actor|
+        begin
+          actor.future(:terminate)
+        rescue Celluloid::DeadActorError
+        end
+      end
+
+      futures.each do |future|
+        begin
+          future.value
+        rescue Celluloid::DeadActorError
+        end
+      end
+    end
   end
 end